jatin_kapoor/requirements.txt
ultralytics>=8.2.0
opencv-python
numpy
pandas
torch
torchvision
streamlit
pyyaml
tqdm
